qa_crowbarsetup: Fix rebootcloud for HA
When doing HA, we can't reboot all nodes in a cluster in parallel as
pacemaker can't figure out by itself what to do when they all boot
together.

So instead, we reboot one by one each node of each cluster. And then, we
reboot all the other nodes in parallel.

This doesn't simulate a power failure for the whole cloud, but that
would be another test case (that won't work for HA).